fix cart fetching for item and total

+++ server/customer_service/src/main/java/com/jump/model/Customer.java
@@ -46,7 +46,7 @@ public class Customer {
 	private String lastName;
 	
 	@Column(name="cart_id")
-	private Long cartId;
+	private Integer cartId;
 	
 	@Transient
 	private ShippingInfo shippingInfo;
